+++ web/src/lib/ui/commands/updateEditLayoutCommand.ts
@@ -1,4 +1,4 @@
-import { type RGLayout } from "~/lib/types/types";
+import { type RGLayout } from "~/lib/types/widget";
 import { useBoundStore } from "../state";
 import { type Command } from "./command";
