+++ web/src/lib/ui/state/widgetLayoutSlice.ts
@@ -1,5 +1,5 @@
 import { type StateCreator } from "zustand";
-import { type RGLayout } from "~/lib/types/types";
+import { type RGLayout } from "~/lib/types/widget";
 
 export interface WidgetLayoutSlice {
   widgetLayout: RGLayout;
