+++ web/src/server/domain/layout/services/updateWidgetLayoutService.ts
@@ -2,7 +2,8 @@ import { BREAKPOINTS_ORDER } from "~/lib/basic/const";
 import { codes } from "~/lib/error/codes";
 import AppError from "~/lib/error/error";
 import Log from "~/lib/log/log";
-import type { RGLayout, ScreenSize } from "~/lib/types/types";
+import type { ScreenSize } from "~/lib/types/types";
+import { type RGLayout } from "~/lib/types/widget";
 import type { AdjustedWidgetLayout } from "../adjustedWidgetLayout";
 
 /**
